+++ ChangeLog.md
@@ -3,8 +3,7 @@
 ## v0.2.0.0
 
 * Introduces `withEphemerides` for bracketed access to the ephemeris file.
-* Introduces "prime" versions of the functions to work on a `MonadFail`.
-* Changes the signature of `calculateCusps` to fail if the underlying library
+* Changes the signature of `calculateCusps` to return a `Left` value if the underlying library
   is unable to calculate the cusps.
 * Improves test coverage.